backend/requirements.txt
fastapi
uvicorn
pydantic[email]
motor
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
requests
tensorflow
numpy
Pillow
certifi
